bonsai: use defaults for uninteresting fields in constructors
Summary:
When constructing `BonsaiChangesetMut`, particularly in tests, there is often a bunch of uninteresting fields which we always specify the default for.  This makes adding new fields to the bonsai changeset tedious as we must update the dozens of locations where this is done.

Replace the majority of them with `..Default::default()` so that we get new default values for free.

In a few places, where we are likely to want to make sure we do the right thing with the new fields, leave them as-is so that the compiler will help us find them.

Reviewed By: andreacampi

Differential Revision: D63395188

fbshipit-source-id: 3ea293a3a8a2499e122a2caa294addb856cfad89