Add scuba logging when {RateLimit,LoadShed}Status::Tracked (2/2)
Summary:
**In this stack:**
This stack adds scuba logging for Tracked Ratelimits and Load Sheds, this will allow us to do a gradual, monitored rollout of ratelimits and load sheds and is specially necessary for default ratelimits defined in https://fb.workplace.com/groups/sourcecontrolteam/permalink/8092969094157699/
**In this diff:**
I pass scuba logger around so stuff actually gets logged. In most intances we can mutate the ctx scuba logger while passing a reference to ensure stuff gets logged properly.
For LFS Server that is not possible but the service is not the goal of the project

Reviewed By: andreacampi

Differential Revision: D62759536

fbshipit-source-id: 0a8626d073a0a427ec0ad97f38769544226b9827