stop proactively destroying NfsdChannel
Summary:
## Background

We saw crashes caused by this proactive destruction of Nfsd channel since there are other operations which are actively using the objects managed by Nfsd. An example is `RpcServer::takeoverStop`.

Moving `channel_` out of scope here triggers the `FsChannelDeleter` to call `destroy()` method of Nfsd3.

## This Diff

We remove this line so we rely on the destruction to happen by itself.

Reviewed By: genevievehelsel

Differential Revision: D63415230

fbshipit-source-id: 2247034c6fd601954401cbe96f0283cd2a842965