Provide hint to the user when the prev command uses large number of steps
Summary:
# Context

Customer feedback - https://fb.workplace.com/groups/2057771867705016/posts/7560618603987694/

# This diff

This diff provides a hint to the user when the prev command uses a large number of steps. The code changes include adding a new hint trigger for the prev command and modifying the hintutil.py file to include the new hint. The hint will be shown to the user if the number of steps is greater than or equal to 1000. The hint suggests using the up command instead.

The current threshold 1000 has been set after experimenting on local Fedora laptop.

In a follow up diff will try to add a config for this.

Reviewed By: zzl0

Differential Revision: D63458118

fbshipit-source-id: 914499ee932cd61a883b366fbdf577dc30e7b2e9