Don't fetch files for public commits
Summary:
We've seen perf issues due to using the file_mods template in `sl log`. We hypothesize this is from being checked out on a large public commit unknowingly, rather than having draft comits with many files changed.

A proper solution involves rewriting things to use `{files}` instead of `{file_mods}`, which may have to fetch and diff manifests (as I understand it). But that's a bit involved.

If our hypothesis is correct, then this fix to skip fetching files for public commits should speed bad cases up.

We'll be able to track this in our analytics by looking at `sl log` durations.

Reviewed By: zzl0

Differential Revision: D63477505

fbshipit-source-id: 4d8112e44dee702ec399fd31bcc140b7cd314cbc