cmdclone: fix Rust clone for legacy repos
Summary: The Rust clone wasn't working with legacy revlog2 repos in tests because it didn't like the "ssh" URL. It was checking for a remote eager repo by the "test" and "eager" schemes, but the revlog2 repo is actually an eager repo under the hood. Fix by delegating completely to EagerRepo::url_to_dir() with no scheme checking ourselves.

Reviewed By: zzl0

Differential Revision: D62680616

fbshipit-source-id: b03462d48f5862df65987598a67831331087feb8