testing: set RUST_BACKTRACE=1
Summary:
This env var is read once and then cached by the Rust std lib. To make it possible to get backtraces in debugruntest in-process mode, let's always set this env var at the beginning.

I had to tweak a few things to avoid splurting out a backtrace in some tests (either making errors not print out Debug, or ignoring the backtrace in test output).

Reviewed By: quark-zju

Differential Revision: D62690570

fbshipit-source-id: 8251d1834ba685b8b2248d0098e818ff2f6a9300