tests: allow cloning eager repos using dummy ssh URL
Summary:
A lot of tests still use the ssh://user@dummy/server URL. This was expecting "server" to be a revlog2 repo. However if "server" was a normal eager repo, the clone would fail.

Allow the clone to work by looking for the "eagerepo" requirement in addition to the "eagercompat" (i.e. revlog2) requirement.

Eventually we should get rid of the dummy SSH url, but for now it makes it a bit easier to update tests if we don't need to also change all the URLs.

Reviewed By: quark-zju

Differential Revision: D62680619

fbshipit-source-id: bd02083d767bf5fa6ea28d5826fb9735e6815b0a