tests: enable Rust clone for "modern" tests
Summary:
Now that the Rust clone works for all kinds of server repos in tests, let's try turning on Rust clone for tests with "modern" configs. This means the tests have things like selective pull turned on, so should be pretty compatible with Rust clone.

I tweaked the `newserver()` function to still eagerly create the server repo and `cd` into it, since that is the contract of `newserver()`.

Reviewed By: quark-zju

Differential Revision: D62680621

fbshipit-source-id: 244f41a09b5dff9c6ca81a0b73d63879e7cff4bb