revisionstore: kill ExtStoredPolicy
Summary:
This handled a complexity with the transition from LFS pointers being stored in the "regular" data store, to them being stored in a separate pointer store. See D22951598.

Anyway, the intention of D22951598 was to use the "ignore" mode when remotefilelog lfs support was enabled, and now that the python lfs extension is gone, we can just bake that in.

I removed the upgrade-lfs-pointers logic from scmstore. I think it was only ever needed for tests since the python extension has been disabled for so long in production. In particular, it handled the case where someone had made a local commit with an lfs file using the python lfs extension (but hadn't pushed the commit to the server). After switching to the remotefilelog lfs implementation, in order to read that (unpushed) lfs file we would need to upgrade the lfs pointer into remotefilelog's separate pointer store.

Reviewed By: quark-zju

Differential Revision: D62714416

fbshipit-source-id: 70bd0485472c97037912dbd27c6e59afeb6cb9ff