pyrevisionstore: fix metadatastore.getsharedmutable
Summary:
This is called in one spot (remotefilelog/fileserverclient.py) and that spot calls `metadatastore.getsharedmutable().prefetch()`. However, the return value of getsharedmutable() did not have a `prefetch` method. This was a mistake I made in D62337845. No tests seem to excercise this, though, and there hasn't been user reports of errors.

Anyway, fix getsharedmutable to return an instance of Self (which includes prefetch). I tweaked the MetadataStore to support copying itself, but removing the local store. I like this pattern better than creating a new shared-only store from scratch because it re-uses the same instance of IndexedLogHgIdStore, so you don't need to worry about stores getting out of sync.

Reviewed By: quark-zju

Differential Revision: D62714412

fbshipit-source-id: 7ecf643af69055efffcd82b892a289eb487c197a