explictly capture reference to underlying context object in lambda to avoid use after move
Summary:
I didn't see this as a crash from the focus week but it appeared on other code as a lint in this path.

This bit of code was refactored in D40349525. As per the comment

https://www.internalfb.com/code/fbsource/[22091fd099e9942bd8d37807f858ad49ed00045d]/fbcode/eden/fs/nfs/Nfsd3.cpp?lines=1985-1987

There was a `auto& contextRef = *context;` which makes a reference to the underlying object, but now we only have a reference to the unique pointer which we dereference.

This code adds the reference to the underlying object back so we unique_ptr move will not invalidate a reference to the unique_ptr itself in the lambda of the inner future

Reviewed By: MichaelCuevas

Differential Revision: D63290488

fbshipit-source-id: 86037016bc5d706983938f8c64b5a360afbb7566