lock: tweak deadlock detection for debugruntest
Summary: Some tests run multiple sl processes at once. Under debugruntest, these are actually threads within a single process. This was tripping the deadlock detection code in lock.py since it only looks at pid. Fix by skipping the check if the current thread isn't the main thread. In general, if there are multiple threads in play, it is "okay" for the current process to acquire a lock when it already owns it.

Reviewed By: quark-zju

Differential Revision: D62680624

fbshipit-source-id: 481b5a740afebc0903241ca7f125e03521b83733