upgrade rollup to 3.29.5
Summary:
The most recent release of react scripts is pulling an older version of rollup, which has a bunch of vulnerabilities associated with it (see associated task).

Since this hasn't been fixed yet ([see this GitHub issue](https://github.com/facebook/create-react-app/issues/13671)), let's just pin it to the first version that fixes this. We cannot add overrides because we are using yarn.

Reviewed By: quark-zju

Differential Revision: D63459489

fbshipit-source-id: b047610af75b89a38b8f5f4e212061114ba098d3