Partially implement a new ping method
Summary:
## This stack

Convert one (non-Megarepo) SCS method to run via async_requests and the async worker.

## This diff

Implement an end-to-end async ping method. This is potentially useful for troubleshooting, as it proves that at least one worker is running correctly. More importantly, this is so simple that it is perfect for the integration test.

This is also an example of a method that takes no "target" argument, i.e. isn't tied to any repo.

This diff only implements enough to submit the request and confirm it is in the queue. The next diff will implement the worker side.

Reviewed By: RajivTS

Differential Revision: D63257080

fbshipit-source-id: c91dff751028834c639b9685c109bafbc243ffb8