Retry unrecognized methods
Summary:
## This stack

Convert one (non-Megarepo) SCS method to run via async_requests and the async worker.

## This diff

So far we've only had two types of errors: either we fail while processing a request, and the client will get a failed response back; or we just crash and hope. That is not great, especially the "hope" part.

We can't make general assumptions that errors from methods are retriable (we might want to do that later). But we do know that some errors are due to clients and servers being out of sync; and retrying those has a good chance of letting us make progress neatly.

Reviewed By: RajivTS

Differential Revision: D63262987

fbshipit-source-id: 7adbd66f5537616a3ad7ce7589abae0ffd77cddc