Finish implementing the new ping method
Summary:
## This stack

Convert one (non-Megarepo) SCS method to run via async_requests and the async worker.

## This diff

Implement an end-to-end async ping method. This is potentially useful for troubleshooting, as it proves that at least one worker is running correctly. More importantly, this is so simple that it is perfect for the integration test.

This is also an example of a method that takes no "target" argument, i.e. isn't tied to any repo.

The previous diff created the foundation; this diff implements the actual response.

Reviewed By: mitrandir77

Differential Revision: D63257793

fbshipit-source-id: 0425eb5fa2562d8a3d127d4aa711130708512df8