git_source_of_truth: allow retrieving entries by repository name
Summary:
We eventually want to get rid of retrieving entries by repository id to
mitigate the situation in the linked task.  As a first step, let's introduce a
method to allow getting the configuration using a repository name so that we
can replace the usage in of repo ids in all call sites with repo names instead.

Reviewed By: RajivTS

Differential Revision: D63633489

fbshipit-source-id: a7db4defbc1b53efcb636ec0382b53965f71469f