User friendly message for commit cloud hook in Mononoke Git
Summary: There is no direct support for commit cloud in Git. The way we work around that problem is by pushing to a commitcloud ref where the push fails but the data gets uploaded to Mononoke. In Metagit, the push failure resulted in a success message describing that the commit cloud upload was successful. In Mononoke, we just a generic hook error message. This diff fixes that.

Differential Revision: D63640744

fbshipit-source-id: d7609124c4d0bbd05f3901e1b2fdf0ff459f6e77