Add a command line flag to control repos filtering
Summary:
## This stack

Complete end to end support for the new async ping method.

## This diff

The current behavior works well as an intermediate state: in particular, it ensure the AOSP Megarepo worker only processes what is intended.

However, we are now adding a command that is not associated to any specific repo; as such, it could be processed by any worker at all.

We could change the query to select "any configured repo or repo is null", but I don't want to introduce any implicit behavior. A flag is easy and very explicit.

Reviewed By: gustavoavena

Differential Revision: D63638130

fbshipit-source-id: e01d268df059a33feb358b352da57bd1302b507a