Fix handling of namespaced refs in Mononoke Git server
Summary: When listing tags in Mononoke Git server, we list them with the `tags/` prefix assuming that all tag bookmarks start with that prefix. However, namspaced refs are an exception that start with `namespaces/`. In production, we saw this problem with AOSP repos like `aosp/platform/external/libcups`. This diff fixes that problem.

Reviewed By: singhsrb

Differential Revision: D63694804

fbshipit-source-id: 581341f1cc7398d9aa26769761c943352d85063a