Add more info about diagnostics confirmation on hover
Summary:
Add a small info icon that you can hover to get more info.

I played with always showing this, with no hover, but it felt like it was in the way.

The content caveats that it's best-effort, and tries to justify why you should take these errors into account.

Reviewed By: madgen

Differential Revision: D63654877

fbshipit-source-id: 3c08fe25d22b31443575f57b4aee8e42713768e1