Log diagnostics error codes and set up ignored diagnostic codes
Summary:
our data was too course, we saw a lot of issues from "arc", but had no way to drill into it.

Let's add error codes and only log the error-level issues. This way, we'll get a clearer picture what errors need to be ignored.

Also, put in the infra for us to ignore certain errors by their source+code. Once we determine which codes are noisy, we can add them to our list of ignores.

Reviewed By: madgen

Differential Revision: D63652771

fbshipit-source-id: 36d892b77292186695928cfeca525cf313faae23