If socket is already deleted in readEof, skip read
Summary:
# Bug Description
When the nfs daemon closes the socket connection abruptly, a readErr is triggered. This causes the RpcServer to delete itself, which then cascades into a readEOF being called after the socket is deleted, leading to a segfault.

# This Diff
Checks if socket is valid when calling readEOF. If not, logs an error then returns.

# Technical Detaiils
Before: Eden crash
After: Mount is unmounted. User gets that popup that happens with stale NFS mounts on eden. Needs to run doctor to fix.

# Discussion Points
I also investigated the alternate option of making sock_ preserved until readEOF.
I chose this method over that one because that was significantly more complicated to implement. Due to how the triggers happen(readErr->RpcServer deleted->mount is unmounted->readEOF is triggered) and the fact that the socket in readEOF is not the same socket as readErr, it is difficult to coordinate the destructors in the right order.

Reviewed By: jdelliot

Differential Revision: D63661824

fbshipit-source-id: fc3d267646509037f176c26cafd3432120f4b6e8