clarify the weirdness in Nfsd3::takeoverStop()
Summary:
It took me a while to understand what was even happening here. This takes some comments from D45584154 and some initial inline comments on earlier version of that diff to hopefully make it a bit clearer what this single line is doing.

I have a feeling somwhere in here is where the crash I'm investigating is located, but the comment is explaining what the desired behavior is (even if thats not how it actually works today, tbd on that)

Reviewed By: chadaustin

Differential Revision: D63668635

fbshipit-source-id: 0f16c57008619b93ce4033f8c5d30190c8fe7b2d