Implement support for shallow push in Mononoke Git server
Summary: The push logic in Mononoke Git server is already structured to fallback on the blobstore backend in case the objects in the packfile sent by the user uses delta bases outside the pack. Shallow Push is essentially a variant of that approach so in a sense we already had support for shallow push but did not have the necessary argument parsing for it. This diff fixes that and adds integration test to show that pushes with and without deltas works in Mononoke Git.

Reviewed By: singhsrb

Differential Revision: D63714705

fbshipit-source-id: b0a9b93ff168f950f1f8a6627bb4e061812b2f85