tests: modernize a few commitcloud tests
Summary:
Use eagerepo and other modern configs.

test-commitcloud-sync.t had an unusual test that I disabled. It was relying on the server ignoring unknown commits in commit graph queries. This is questionable behavior we are exploring separately for Mononoke, but for now we did not want to add that behavior to eagerepo to satisfy this test.

Reviewed By: quark-zju

Differential Revision: D62680617

fbshipit-source-id: d1ea5c14a6108f7684263025967fd79c5fb71b15