release reliability: add install.toml for DevFeature eden
Summary:
## Background
We want to add Eden FBPKG to DevFeature to be live-installed to OnDemand. Because FBPKG is less smart than RPM, we have to explicitly tell DevFeature how to create symblinks for the binaries and static files. DevFeature will query "install.toml" file located in the root folder of the FBPKG.

##This Diff
We add a file which defines the mapping between the targeted symblink and the path of files within the FBPKG package itself.
Also, we need to include this file into FBPKG package.

After this diff is landed, we should be able to test D62906258.

Reviewed By: kmancini

Differential Revision: D62907962

fbshipit-source-id: 41b4b8e3bdea217ca89a37e8562d31039a15d057