rename Metadata -> Aux Data
Summary:
# Context

We're very inconsistent with the way we refer to extra information that's attached to blobs or trees. In some places we use Metadata, and in others we use "aux data". For example, in Eden we have used BlobMetadata and TreeMetadata to refer to this extra data. In Sapling, we also refer to this extra data as File and Tree Aux Data. In Sapling, "metadata" is used to refer to an entirely different type of data. And finally, Mononoke uses Metadata in most places. Lastly, EdenFS uses "Metadata" in a number of places, the most confusing being InodeMetadata which refers to the information that's returned by a `stat` call (ex. mtime, ctime, blk_size, etc).

Given all the above information, during the RE CAS focus room we decided that unifying the clients on "Aux data" would be the best path forward.

# This Diff

This diff migrates as many references from {Tree,Blob}Metadata -> {Tree,Blob}AuxData as possible. A few key places are left out:

1) Any thrift definitions cannot be changed. This is unfortunate, but we can possibly use aliases? That decision is open for discussion.
2) Any ODS counter names (i.e. the actual strings used to differentiate counters when querying ODS) are left as "metadata". I'd like to eventually change these, but we're stuck for now. We might be able to start duplicating some of those counters and then slowly migrate after ~90 days (or whatever our retention is).
3) The Markdown files for Eden stats are left unchanged due to #2 above.
4) The SerializedBlobMetadata fuzzing target and file are left unchanged. There's some weird security setup that I can't figure out, and it prevents me from modifying the target name or the sources of the target.

As far as I can tell, everything else is changed. Please let me know if you think I missed anything.

Reviewed By: jdelliot

Differential Revision: D63152972

fbshipit-source-id: 91d935bdb92a7d0f59db5fbbd8c54572b3585a1d