rename tree blake3 -> digest hash
Summary:
# Context

We previously overloaded the Blake3 attribute to return tree digest hashes to users. Therefore, it made sense to use "blake3" terminology throughout our structures. We then found out that overloading was a bad user experience as users were not expecting to get tree blake3s (i.e. digest hashes) along with file blake3s. This caused performance regressions in some cases. Therefore, we introduced a new attribute DIGEST_HASH which can be used to fetch digest hashes for trees and files.

With the introduction of the new attribute, we should make sure that we name the tree metadata attributes accordingly throughout the code.

# This diff

Changes a bunch of TreeMetadata references to use "digest hash" instead of "blake3"

Reviewed By: kavehahmadi60

Differential Revision: D62587883

fbshipit-source-id: 28dbd5cb11f8d9dc7b6623c19fd13daad7a631d1