Ensure that repo and method gets populated in Mononoke Git Scuba
Summary: Sometimes a call to Mononoke Git server can fail even before all the contents of the request have been parsed. We saw this in production with few aosp repo and chromium/src repos. Lets ensure that we always log the repo name and the method for every incoming request so we get meaningful logs.

Reviewed By: YousefSalama

Differential Revision: D63756774

fbshipit-source-id: 6b15264afa44001315dce81ef06c469420799ecd