Final SoT validation in Mononoke Git Server before committing push
Summary:
The `git push` process in Mononoke Git Server involves multiple steps and each of those steps can take a while if the push is big or for a heavy repo. This forces us to be very conservative when switching the SoT from Mononoke to Metagit since we are always concerned about inflight pushes. With this diff, we greatly reduce the ambiguity period by additionally checking immediately before accepting the push.

This will allow the SoT switch to happen in a minute or two instead of waiting for 20 minutes as is the case today.

Reviewed By: singhsrb, lmvasquezg

Differential Revision: D63758000

fbshipit-source-id: 1a804a8ce96ba650761bd0ac77dc2ac98cdb68d2