rename HiveLogger to ScribeLogger
Summary: Addresses a longstanding TODO to give this class a more reasonable name. `HiveLogger` was an implementation detail that didn't end up being realized, but at the time, there was a different `ScribeLogger` class anyways. That other `ScribeLogger` class was removed, so this one can be renamed to the more fitting `ScribeLogger` name now.

Reviewed By: kavehahmadi60

Differential Revision: D63711173

fbshipit-source-id: 947b9349e9d7e0dfa939c06283ca1c14fe87d95d