git_server_push_poller: add timeout for remote git operations
Summary:
The investigations in the linked task revealed that some of the poller loop
iterations were taking a long time based on
https://fburl.com/scuba/git_server_push_poller/qfm53voa.

This can happen if all the tasks serving a repository are restarting and will
be fixed by D63705001.

The improvement here is that we know that that remote git operations in the
poller should not take more than a couple of seconds. Therefore, let's just
fail the repository processing after a timeout so that the poller can report
it's doing some work and not alert as it did this morning.

For now, the timeout is chosen to be a fairly aggressive 5 seconds which should
suffice for the current repositories because the operation does not take more
than a couple of seconds for now. However, I am also exposing a JK to increase
the timeout in case it ends up being the bottleneck.

Reviewed By: RajivTS

Differential Revision: D63754597

fbshipit-source-id: 88a0681edb5894890403b3a99c1bec692cf2acd3