Update strip_glog helper
Summary:
## This stack
Implements and tests the logic on `git pushrebase` to handle pushes when the source of truth is in the large repo (e.g. fbsource).

In the next few diffs, I'll be adding support and tests for pushing to non-master branches using `git pushrebase` when the SoT is in the large repo.

# This diff
The current regex doesn't match logs with TRACE verbosity when using the tracing library (since D63700050).

Reviewed By: lmvasquezg

Differential Revision: D63758124

fbshipit-source-id: 15e678659c13acc4064a48a6458ee90555ee9b59