Parameterize sync folder in test helpers
Summary:
## This stack
Implements and tests the logic on `git pushrebase` to handle pushes when the source of truth is in the large repo (e.g. fbsource).

In the next few diffs, I'll be adding support and tests for pushing to non-master branches using `git pushrebase` when the SoT is in the large repo.

# This diff
In the next few diffs I'll make changes and add a test to support pushing to non-master branches with `git-pushrebase` when the SoT is in fbsource.

For now, it will work under the assumption that the bookmark prefix matches the git repo name, so let's parameterize that and use it for the small repo sync directory and the bookmark prefix, which is what we do in prod.

Reviewed By: lmvasquezg

Differential Revision: D63758126

fbshipit-source-id: 76bf42a6556dbcb505017cb00038f28d08da028d