Refactor the existing computation
Summary:
## This stack

Many Git repos have little activity, so the objects count may not be a good proxy for the cost to SCS of hosting their shard. Make it possible to override or provide a scaling factor.

## This diff

Move the entire computation to a single function so it's easier to update.

Also add unit tests, which bring prompted me some additional small cleanup.

Reviewed By: gustavoavena

Differential Revision: D63762748

fbshipit-source-id: a69e020f69efd8f42c05af8cbd4c6f0ca51a2fb1