Fix context lines not loading if cwd is not repo root
Summary:
If your cwd is not the repo root, then our call to `sl cat` when producing extra lines will be incorrect, because `sl cat` will interpret the file path as being cwd-relative.

instead, we can construct an absolute path when calling `cat` for the context lines. The signature for `cat` asks for an Absolute path, so this was just a mistake.

Reviewed By: sggutier

Differential Revision: D63778219

fbshipit-source-id: 8646ca79e71a372bf85f114c6e206f6c3837295b