add eden_fsck to list of targets
Summary:
# This diff

makes buck builds set the EDENFS_FSCK env variable so that the source built fsck binary is locatable in tests and dev CLI builds

# Context

The `eden fsck` command invokes the C++ fsck binary to perform file system checking. The issue is that it checks for the EDENFS_FSCK env variable to determine if it should use the production fsck or a source built fsck. EDENFS_FSCK is never actually checked anywhere in our tests, so we always end up using the system installed (production) fsck binary. This causes tests to fail if you make changes to the system fsck.

Reviewed By: jdelliot

Differential Revision: D63849445

fbshipit-source-id: 36ec29fbe73014e3a44687644194210bd9236e20