Update scs repo-info command documentation
Summary:
## This stack
Implements the last remaining component for `git pushrebase` to work when the SoT is switched to the large repo: a way to check with SCS if the source of truth for a given repo is in another large repo and this repo's name.

## This diff
Just updates doc comments that are wrong because were probably copied from the `info` command when the command was created.

Reviewed By: lmvasquezg

Differential Revision: D63882399

fbshipit-source-id: 79ad0196552d392da8aff091058c84304c9f1429