Extend scsc repo_info command to return push redirection repo name
Summary:
## This stack
Implements the last remaining component for `git pushrebase` to work when the SoT is switched to the large repo: a way to check with SCS if the source of truth for a given repo is in another large repo and this repo's name.

## This diff
Extends the `repo-info` command to return the name of the large repo that is the source of truth for a small repo.

Reviewed By: lmvasquezg

Differential Revision: D63882400

fbshipit-source-id: 5285e2ae9d59d82a53e634dd1a7bd8755f909241