tests: disable cas for Sapling .t tests
Summary:
Diff D63730014 was intended to excluding sapling tests for CAS.
however, the hgbuild tests use the real linux built sl, so they
are trying to use the real cas client (thanks Muir)

We also need to un-disable that for Mononoke integration tests, since
we want to use cas client there. For sapling .t tests, we are using eagerepo
cas client (which is not disabled by cas.disable)

We are planning to make the lack of repo name propagate None for the
cas client instead of an error, similar to how eden_api works.

Reviewed By: muirdm

Differential Revision: D63921681

fbshipit-source-id: 211512f9c0fe5699ac1f4059cb18cbe999d205f8