Configuration to prevent deep nested stream in GitImport
Summary:
AOSP's MGS tool stopped working when we landed D58675665 due to streams not making progress at all. To fix this, we had to rollback. However, the changes in the diff contain useful changes that we want to make across Mononoke codebase. This diff introduces a flag that avoids the hanging stream problem by collecting the stream and then processing it.

Gitimport, Live Git Sync and Mononoke Git server will keep using it the old way while MGS can choose to collect the stream by flipping the flag.

Reviewed By: mzr

Differential Revision: D63987060

fbshipit-source-id: e5f8391dc7284efc0d7cec103435dea37bdc9749