Rename setEditedCommitMessage to setEditedField
Summary: `setEditedCommitMessage` is renamed to `setEditedField` in the `CommitInfoTextArea` component, to be more clear. This is because `commit message` can otherwise mean `summary` or all diff authoring fields combined. In this case, only one diff authoring field is being referred to, and it is not necessarily `summary`. For example, the `test plan` field can be referred to.

Reviewed By: evangrayk

Differential Revision: D64000372

fbshipit-source-id: c453007d3057b036e4658e3ccf58afefe1623a85