Add isResolved to DiffComments
Summary: Add the type for if a comment is resolved and fetch it for Phabricator Diff Comments. `null` means N/A, for example for comment replies. `true`/`false` mean we should render the status in the UI as either Resolved or Unresolved, and possibly add actions to mark it resolved.

Reviewed By: quark-zju

Differential Revision: D63994221

fbshipit-source-id: 141cfabffb58fc2c3a8c97b0d80df4254e9fb565