Surface resolved status in comments preview
Summary:
Now that we have the comment resolved status, add a small indicator to comments in the preview.

This should ideally actually be a button, not in indicator. On Phabricator, it says "resolve" as a clickable button instead of "unresolved". This read-only indicator is an ok place to start. We can make this clickable in the future.

Reviewed By: alichtman

Differential Revision: D63994332

fbshipit-source-id: 77894ed8953efc8925b56cc3dd436de2198bad1f