Add auto-generating diff test plan feature in ISL behind GK
Summary: The feature to auto-generate a diff test plan is added in ISL, behind the GK `inject_test_context_for_autogenerated_test_plans`. Note that the GK check is carried out in the `TextArea` file because if it is carried out in the `InternalCommitMessageFieldsSchema` file instead, then there would be a circular dependency.

Reviewed By: evangrayk

Differential Revision: D64006433

fbshipit-source-id: 8f03ab51b59d7250f630dc8e7d3fc337e1aea358