commit cloud: properly convert timestamps on first insertion
Summary:
T202321515 mentions some errors in version insertion queries that may be causing data loss. `timestamp` is a column with type `timestamp`, therefore the issue is not with `current_tiemstamp` since this is the default for this type of column, it also doesn't match the quantity of errors (~500 for a week).

My guess is that when we insert the first version of a workspace the types are not matching since we're sending an i64 and the query expects a timestamp, so fixing that.

Reviewed By: markbt

Differential Revision: D64004093

fbshipit-source-id: 44093b26ef5abfe7bf8a02db43a4a4ac4fb51067