Enable creation of Eden API client based on paths.default
Summary:
Currently, when creating the EdenAPI client in Sapling, we rely on the `edenapi.url` config setting for getting the base URL for the HTTP client. However, if we ignore wireproto the base URL can also be constructed based on the `paths.default` configuration using a custom scheme. This diff adds support for creation of the Eden API client using the `paths.default` config if:
- Its available
- Maps to a scheme dedicated for Git repos in Sapling

Reviewed By: quark-zju

Differential Revision: D64011460

fbshipit-source-id: 06af09b66efa86165c0bd1177be2fdeb75a7aee3