Log validation failures encountered during Git Push
Summary: We log system errors that we encounter during `git push` but if the push fails due to validation or logic or hook errors, then we return success and nothing gets logged in scuba. When debugging potential misconfigurations in production, it is useful to have access to logs even for such validation error cases. This diff does that.

Reviewed By: singhsrb

Differential Revision: D64018286

fbshipit-source-id: d04e82f5f510361c6bf5d1e189a21cfd7ef37af3