initial implementation of PrefixTree
Summary:
`TrieMap` is used frequently in derived data code but it's not memory efficient as it needs `O(number of distinct prefixes)` pointers. I'm implementing a prefix tree alternative to it so that we will need only `O(elements)` pointers. Using `SortedVectorTrieMap` everywhere instead is another alternative that would be interesting to benchmark against.

This diff implements inserting key-values pairs into the prefix tree, and iterating over all of its key-values pairs.

Differential Revision: D63541165

fbshipit-source-id: 034c32b69ffe7db2a722c6eb7f59c2001aa1b92b