Handle redirection symlink in windows: Keep track of redirection target in checkout config
Summary:
# Background
We have seen tools like Unity messing up with already existing redirection symlink by deleting and/or creating a folder/file instead of symlink.

# Solution
When a create/delete operation is triggered for a redirected path, handle it in a way such that reidrection stay intact.

# This change
Keep track of redirection target path in checkout config so that it can be used while handling projected-fs notification for these operations triggered by user.

Differential Revision: D63945643

fbshipit-source-id: 27a296ad906fb87832fdeffb38295a6f87313c50