Extract fetchAdditionalLines in comparison context
Summary:
The `useFetchLines` hook is problematic for reusing the split diff view, because it transitively imports a lot of ISL code.

To extract the split diff view, we need to make each piece reusable and easily importable.

To start this, let's extract `fetchAdditionalLines` so the serverApi part is in ComparisonView and not SplitDiffView. We're not quite ready to move all of useFetchLines yet, but can do that next.

Reviewed By: quark-zju

Differential Revision: D63778220

fbshipit-source-id: 29edaecbc6ee1f20346f75d8a2f2e8ad31dde4b5