fix more ambiguous calls
Summary:
# This diff

Prefixes our call to format_to with `fmt::` to remove the ambiguous function call error during compilation.

# Context

This is the exact same issue we hit in D63902777, which fixed references within Watchman. To recap:

C++ 20 added std::format_to, which causes bare format_to calls to be ambiguous. We can easily fix the issue by qualifying the calls with fmt:: as done in this diff.

Reviewed By: kavehahmadi60

Differential Revision: D64085858

fbshipit-source-id: 5f1536baa5cfd8dc31ae8c5eba7e1b4e029138f3