CachingTreeStore: add cache_with_key method
Summary:
# This diff

Adds a method to the CachingTreeStore that allows adding entries to *only* the cache and not the underlying TreeStore. This will allow us to avoid persisting materialized trees since they are not super likely to be reused in the future.

# Context

It's currently not possible to get TreeAuxData for trees that are materialized within an Eden mount. This is because those trees are not exposed to Sapling until a commit is made (or status is run, but status doesn't persist the materialized trees). This comes with big downsides, as TreeAuxData for materialized trees would be useful for many workflows.

EdenFS like to support requesting TreeAuxData for materialized trees. In the past, we've had Eden derive aux data for files on demand, but that leads to duplication of logic to calculate aux data. We want to avoid teaching many different parts of the source control stack how to derive TreeAuxData to avoid making the same mistake as before.

Therefore, the TreeAuxData calculation logic will live within Sapling, and Eden will call into Sapling to derive TreeAuxData for materialized trees. The following is a rough outline on how this will work:

1) EdenFS will notice that Sapling is missing a TreeAuxData entry for a materialized file
2) EdenFS will call into Rust to request that Sapling calculate TreeAuxData for the materialized tree.
3) Sapling will call the getScmStatusV2 endpoint to get a list of materialized files/dirs.
4) For each materialized file/directory, Sapling will create sapling objects for them. TBD on how Sapling will read the modified files/dirs (either through the filesystem or through a special thrift end point that we expose to Sapling)
5) Sapling will add each new tree to a CachingTreeStore (this ensures the new objects are stored in memory and not actually persisted in the underlying stores)
6) Sapling will call into a shared library that calculates digest hashes for objects using a provided store that contains all the trees/files
7) Sapling will return the TreeAuxData to Eden and possibly cache the result

Reviewed By: quark-zju

Differential Revision: D63915057

fbshipit-source-id: 1e1cd8196424456be5a7daee575d62fcc6949cf3